fix_3. Score is 100

+++ tasks/tutorial/aplusb/aplusb.hpp
@@ -1,5 +1,5 @@
 #include <iostream>
 
-int Sum(int a, int b) {
+long long Sum(int a, int b) {
     return a + b;
 }
